feat(vehicle): add loading component for vehicle page fix(vehicle): add delay to simulate loading data in vehicle page
The `loading.tsx` file is added to the `src/app/vehicle/[id]` directory. This file contains a `Loading` component that displays a loading animation while the data is being fetched.

In the `page.tsx` file, a `sleep` function is added to simulate a delay of 10 seconds before rendering the page. This is done to mimic the loading of data. The `Tabs` component is also updated to include a `defaultValue` prop to set the initial active tab.